requirements.txt
youtube-transcript-api==0.6.1
requests==2.31.0
python-dotenv==1.0.0 